440: Refactoring functional tests to support reusing test logic across multiple versions which are functionally identical
Versions 3.1.8 and 3.1.9 are functionally equivalent for all the schemas (except VRPs which we currently don't have tests for). We would like to be able to run tests against the different versions that we support to verify that the controllers behave identically.

Refactored tests to allow the version to be plugged in, which enables us to execute the same logic against different API endpoints.

GetAccountBalancesTest demonstrates how simple test cases can be refactored, extracting a single test method which the version can be plugged into. This works as both of the methods are identical except for the urls.

CreateDomesticPaymentsConsentsv3_1_8Impl, CreateDomesticPaymentsConsentsv3_1_8Test and CreateDomesticPaymentsConsentsv3_1_9Test demonstrate a pattern for more complex test cases where there are multiple test cases that we want to reuse. Here we extract a class with the test implementations and have separate versions specific Test classes which invoke the test methods from JUnit.

CreateDomesticPaymentsConsentsTest class has marked as deprecated and only contains tests for versions prior to 3.1.8 - as it was deemed too much effort to refactor these when it is likely that we will drop support for older versions in the future.

Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/440