494: Fixing payment consent tests to send a detached JWS signature in the request header
There was a bug in the consent related tests, which was causing a JWS with payload to be sent in the x-jws-signature header, when a detached JWS is required.

Refactored the code relating to creating detached JWS signatures and interacting with the Payment API. New PaymentApiClient should be used instead of the existing PaymentRS (which has been deprecated). Migrated payment consent tests to use the PaymentApiClient.

Additional cleanup related to getting access tokens has been made by shifting the logic shared by the AccountRS and PaymentRS into the Tpp class, so that there is one set of logic for getting client_credential based AccessTokens

Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/494