980: Improving assertions of International Payments tests to verify that ExchangeRateInformation is generated in a consent response when it was omitted from the request
The field: consentRequest.data.initiation.exchangeRateInformation is optional. The tests which submit consents with only the mandatory fields will not be setting this field.

Improving such tests to verify that the consentResponse.data.exchangeRateInformation is generated by the ASPSP, and that the initiation.exchangeRateInformation is not changed in the response.

This applies to International Payments and International Scheduled Payments.

https://github.com/SecureApiGateway/SecureApiGateway/issues/980