1012: Fixing Domestic Payment Consents to work with RS impl of Consent API
The Domestic Payment Consent API is now being implemented by controllers in the RS (previously it was being implemented in IG). Changes have been required to the tests to deal with different error messages.

Charges for Domestic Payments have been removed in the RS impl, as this is more representative of typical Domestic Payment transactions for a UK Bank.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1012